criando classe AlertManager
